Methods to make a request to the API were implemented #19
Make request was made to check if there's a token and to create the url to create the request. The url is gotten with the getUrl method which help us build the url with the base API url and the query parameters. Once the request has been created, the method call is used to make the call to the Spotify API with an Http Client. Inside this method we'll be able to manage our objects and add them to the results displayed on the recycler view.