Clear and Add methods were created #19
These methods were created in case you need to clear your list of objects or add objects to the list. The clearAll method it also asks if you want the adapter to be notified about clearing the list in order to make disappear the objects being displayed on the screen.