Wait for the API response #25
To return the list of songs i have to make sure the API has already give an answer. In order to do that, i created a new thread to execute the callAlbum method, which is the one that makes the request and handle the results. Once i make the instruction to execute the runnable, i await for its termination in order to return the list of songs.