Get Album songs method (Songs from each album) #25
Once i have a list of albums that matches the query, i have to take the songs from each album and store them in a list. To get the songs from each album i call the createAlbumSongs method which returns a list of songs. At the end i return a list with all of the songs from all of the albums .