API search song request was modified #23
In order to get all of the songs that matches the query, i get the call to the API from the Spotify class and handle the response inside the activity. With help of the createSongs method from the Spotify class, i convert the response into song objects and once i have the response i notify the adapter to display the songs on the screen.