API search album songs request modified #23
Instead of waiting for the API response, i handle the album search response inside the activity. First i get the call to get the albums that matches the query and with help of the createAlbumForSongs method from the Spotify class i can send the list of albums and i get back the list of songs of each album on the response. Once i have that i notify the adapter to display the songs on the recycler view.