Get album songs request #30
The request is made through the Spotify object and the API response is handle from the fragment with the help of the createAlbumSongs method on the Spotify class which returns a list of song objects that belong to the album. Once the response is gotten the adapter gets notify and the objects display on the recycler view.