Song object from song detail #32
If the add button was clicked from the song detail, the song object from the compose activity is set to the song object from the intent in order to show the details of the song picked. If the compose activity is being open from any other activity or fragment then the song object is not modified, this just happens if it's open from the song detail.