Adapter gets notified #39
Once the list of facts that the user has made has been gotten from the database, i clear the list of the adapter and add all the new fact objects. This is done in order to get get the older and the newest facts without they repeating on the list (that's why the list gets cleared first).