OnActivityResult method #40
Once the user has picked an image from their gallery, the activity handles the answer on this method. The loadFromUri method returns a bitmap gotten from the uri. That bitmat is loaded into the image view. To save the image as a parse file (in order to upload it to the database) we have to convert the bitmat to a byteArray.