Load next data method was created #46
On this method a new request is made with the offset parameter as the page gotten from the scroll listener. Once a new set of data is gotten, the adapter gets notified and the recycler view is updated. I recognize if the tab is the song tab or if it is the album tab in order to know which adapter i have to notify.