Load next data method was created #47
Just like on the search fragment, i have to check if the user is on the song tab or the album tab to know which request i have to made. Once i know that  i make the request with the page gotten from the scroll listener and finally i notify the adapter.