Click listener interface was created #49
This interface will allow me to control the function of the on click listener that i indicate. In this case i'll use it to control the delete button which will only appear on the profile fragment but i have to implement the interface on every class where i use a fact adapter.