queryLikedFacts #50
If the user is on the Likes tab then the queryLikedFacts method gets called which makes a request to the database to get back the facts in which their object id match with one of the objects id stored on the factsLiked array which is an attribute on the user model. Once i have the facts liked by the user, i notify the adapter.