Like adapter was created #51
Since the like tab and the my facts tab make different requests, i have to handle liking a fact differently. When a user is on the my facts tab and like or dislike a fact it only has to update the data on the database, but when a user is on the likes tab i have to update the data base and the screen since i'll have to add or remove facts from the recycler view.