For statement on the MyFacts Tab #51
If the user is on the MyFacts tab, i go through the entire likedFacts array (Which contains the objects id from the facts the user has liked) and check for each object if their objectId equals the objectId of the fact on the position, if it does it means that fact has already been liked which means the user in unliking the fact so i remove the objectId on the position where it was find and save the new list on the database. If it's not equal the that means the user is liking the fact so i add the objectId to the list of liked facts and save it on the database.