For statement on the Likes tab #51
Just like on the MyFacts tab i check if the fact has already been liked, if it has i remove it, upload the new list to the database and make the call with the queryLikedFacts in order to disappear that fact from the screen. If it hasn't been liked i add it to the likedFacts list, upload the new list to the database and call tge queryLikedFacts in order to update the recycler view.