onLikeClick method was implemented on the home fragment #51
All of the onLikeClick methods do the same, the only thing that changes is the array where the facts are stored. In this case i have a list of all of the facts stored on the database, still we only need the object id in order to update the database correctly.