OnRefresh method functionality #56
Since the recycler is used for two different things i have to first check in which tab the user is on, once i know that i can call the corresponding query method. If the user is on the MyFacts tab i call the queryMyFacts method, if the user is on the likes tab i call the queryLikedFacts method.