Modified queryFacts parameters #58
Since i need the page to know which set of objects i'm going to ask the database for i have to pass the page number and a boolean which indicates if i want to clear the list of the adapter or not. This is to avoid clearing all of the objects when i just need to load more without clearing the ones i already had.