query methods are called with the new parameters #59
These are the calls i make where i need the data to display from the first set of objects and that's why page is equal to 0 and clear is equal to true, because is used more when you want to update the data. In this case a really clear example is when we like or dislike something being on the likes tab, i need to update the data to delete a fact from the recycler view when a user dislikes a fact from there.  and thats why i need to make a call from the first set of objects.