onLoadMore method functionality #60
Since the set of objects grow 10 every time the page variable will be equivalent to the multiplication of the page by 10. If it's the first page the database will give us from the 10th object to the 19th and so goes on with the next pages. Once i have the page number i call the loadNextData method.