Album detail case #70
If the string doesn’t match with any of the previous cases that means the petition came from the only other option which is the album fragment. This means this fragment is being displayed on the child fragment of the album detail, which also means that to return we need to get the child fragment manager of the parent fragment and remove it.