Remove current user from the results #72
To do this i go through the result list and check if the object id from the user on the position is equal to the object id of the current user. This is made in order to don't show the current user on the results because that wouldn’t really make sense.