First fact liked #73
If the list size was 0 then that means that the facts is being liked so it's added to the list directly (it wouldn't work with the for statement because it wouldn't make a single iteration). The number of likes o the fact is updated and the list of facts liked by the user is updated too.