Fact being liked #73
If the list was already checked and the fact wasn't found, that means the fact is being liked so i update the number of likes on the fact and update the fact on the database. Once that's done i update the list of facts liked by the current user and save it on the database.