Pass user object #73
This is to return to the user searched profile when the close button is clicked on the song fragment. For example, if i searched for nate's profile (someone random) and i access to a song detail from there, when the i clicked the close button i'm going to be redirected to nate's profile again.