Number of facts is gotten from the database now #74
Since the fact number was gotten from the adapter item count method, the number was always updating because if the user has more than 10 facts, then we have to keep fetching the data from the database. This means that if a user has 20 facts, the number of facts used to appear as 10 first because the data wasn't loaded yet. To solve this i put the number of facts as an attribute for the user.