The user is on the list case #77
If the user was found on the list that means the user is being unfollowed by the current user. RemoveAt is a variable which will be used on the addFriend method to know in which position to remove the object. If this is the case, the method returns true because it means the current user follows this user.