Get expired hour from shared preferences #69
This number indicates the hour where the token will be expired on minutes (including the minutes of the time). For example if the time at where the token was generated is 15:30, the hour stored will be 15 * 60 = 900 minutes + 30 minutes = 930 minutes, so the token will expire at 990. If we calculate this:  990 - 30 minutes = 960 /60 = 16, which means the time is 16:30 so the token has expired and new one has to be generated.