Receives shared preferences instead of context #69
This because i used to get an error about the context being null. I don't fully understand but i think it may have to do with asynchronous tasks, like maybe when i created the Spotify object on an activity and the Spotify class called this method the context was still null or just not valid to call at that moment because i read that sometimes context can only be called when the activity is created but i'm not sure. I thought that since i already store the shared preferences on the Spotify class it will be easier for me to pass it instead of creating it again.