Check if friend method reused #86
This method goes through the entire array of the user friends and see if the user clicked is on the list, if it is then the method returns true, if it's not on the list then the method returns false. If the friends list is empty the it returns false.