Unfriend user #86
If the method returned true, that means the current user is clicking the button to unfriend the user so i update the number of friends and the list of friends of the current user. Once the user is updated on the database i notify the adapter in order to change the follow button image.