For statement to go through the array  #85
Checks if the user at the position on the friends list is the same user clicked, if it is then the method returns true because that means the current user follows the clicked user.  If the list has been checked and the user is not found, the method returns false.