QueryFriends method with page updated #85
This method will be call every time a new page is needed. The clear variable is false in order to not delete items before adding the new ones, in this way the list of friends is complete. setSkip skips the first n objects, that's why i used it with the page variable.